+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 * Add inclusive/exclusive options to clj-toolbox.colls/take-range
 * Convert some extraneous macros in clj-toolbox.test-utils to be a function instead
+* Add clj-toolbox.files/copy
+* Add clj-toolbox.files/backup!
+* Add clj-toolbox.files/home
 
 ## [2.0.0] — 2025-04-30
 * Remove deprecated functions from clj-toolbox.files